+++ src/App.js
@@ -1,20 +1,13 @@
 import React from 'react';
-
-const todoList = [
-  { id: 1, title: 'Complete assigment' },
-  { id: 2, title: 'Play with kids' },
-  { id: 3, title: 'Watch TV' },
-];
+import TodoList from './TodoList';
+import AddTodoForm from './AddTodoForm';
 
 function App() {
   return (
     <>
       <h1>Todo list</h1>
-      <ul>
-        {todoList.map((item) => {
-          return <li key={item.id}>{item.title}</li>;
-        })}
-      </ul>
+      <AddTodoForm />
+      <TodoList />
     </>
   );
 }
